Strokes are drawn onto seperate canvas and then applied to the main one
In order to have pen pressure, the one giant stroke needs to be a large
amount of smaller strokes instead. This is because we can't change
values in the middle of a stroke. In this method, first the stroke is
drawn onto a canvas, which then has a rectangle applied on top of it
which will subtract some opacity