Undos/Redos are now synced across clients
Changes now have 'Owner' and 'Active' fields. Instead of tracking the
most recent change, when an undo happens, we change the most recent
active change by a user to be inactive. With Redo, we find the oldest
inactive change and make it active. These are all synced using basic
events and data.

Has not been tested with merging/removing layers